pyproject.toml
[project]
name = "qwen-siglip"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=3.36.1",
    "pillow>=12.0.0",
    "protobuf>=6.33.1",
    "sentencepiece>=0.2.1",
    "tensorboard>=2.20.0",
    "torch>=2.9.1",
    "transformers>=4.57.1",
]
